requirements.txt
streamlit==1.37.0
streamlit-authenticator==0.4.1  
streamlit-navigation-bar==3.3.0 
plotly==5.4.0
pandas>=2.2.2
numpy>=1.26.0,<2.0.0
seaborn==0.12.2
openpyxl==3.1.2
alt==0.0.4
folium==0.14.0
geopy==2.3.0
streamlit_folium==0.12.0
streamlit-on-Hover-tabs==1.0.1
Faker==33.1.0
sqlalchemy==1.4.46
polars==1.8.2
psycopg2-binary==2.9.10
duckdb==1.1.3
gspread==6.1.4
oauth2client==4.1.3
PyMuPDF==1.22.5
langchain==0.0.220
docx2txt==0.8
openai==0.27.8
transformers==4.46.3
tldextract
python-whois
tldextract
requests
